+++ lib/automated_testing_framework_plugin_firebase.dart
@@ -1,3 +1,5 @@
+export 'src/driver/firebase_test_driver.dart';
+
 export 'src/firebase_test_store.dart';
 
 export 'src/forms/assert_firebase_value_form.dart';